+++ CHANGELOG.md
@@ -7,6 +7,7 @@ for future and past milestones.
 
 New features:
 
+* Added support for dropping files on signal/image list widget
 * Added "Computing parameters" group box to show last result input parameters
 * Added "Copy titles to clipboard" feature in "Edit" menu
 * Image processing: